+++ reffect_core/src/context/resource.rs
@@ -2,22 +2,22 @@
 #[derive(Debug, Clone)]
 pub struct Resource {
     /// Current amount.
-    pub current: u32, // TODO: as f32?
+    pub current: f32,
 
     /// Maximum amount.
-    pub max: u32,
+    pub max: f32,
 }
 
 impl Resource {
     /// Creates an empty resource.
     #[inline]
     pub const fn empty() -> Self {
-        Self::new(0, 0)
+        Self::new(0.0, 0.0)
     }
 
     /// Creates a resources with the given values.
     #[inline]
-    pub const fn new(current: u32, max: u32) -> Self {
+    pub const fn new(current: f32, max: f32) -> Self {
         Self { current, max }
     }
 }
